pyproject.toml
[tool.poetry]
name = "lostcat"
version = "0.1.0"
description = ""
authors = [ "Joel Sleppy <jdsleppy@gmail.com>" ]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.4"
django-extensions = "^3.1.3"
Pillow = "^8.2.0"
django-model-utils = "^4.1.1"
gunicorn = "^20.1.0"
fabric = "^2.6.0"
sentry-sdk = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-django = "^4.3.0"
factory-boy = "^3.2.0"
black = "^21.5b2"
django-debug-toolbar = "^3.2.1"

[build-system]
requires = [ "poetry-core>=1.0.0" ]
build-backend = "poetry.core.masonry.api"
